requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
anthropic>=0.25.0
pydantic>=2.0.0
python-dotenv>=1.0.0
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.9
